feat: SolutionFixture 수정
고정된 필드는 유지하고 아닌 필드들은 메소드 파라미터로 받도록 수정

Related to : #9
